requirements.txt

brotlipy==0.7.0
celery==5.3.6

click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0

datashape==0.5.4
Django==4.1.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-silk==5.0.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
